pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "AI Agent Workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.1.1",
    "crewai>=1.1.0",
    "groq>=0.33.0",
    "langchain>=1.0.2",
    "langchain-core>=1.0.0",
    "langchain-groq>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlite-utils>=3.38",
    "tavily-python>=0.7.12",
]
